+++ app/api/concerns/api_configuration.rb
@@ -1,0 +1,14 @@
+module ApiConfiguration
+  extend ActiveSupport::Concern
+
+  included do
+    cascade true
+
+    # include ErrorsHandling
+
+    format :json
+    default_format :json
+
+    helpers ApiResponseHelper
+  end
+end