+++ app/api/helpers/api_response_helper.rb
@@ -1,0 +1,5 @@
+module ApiResponseHelper
+  def say_ok
+    status :no_content
+  end
+end