+++ config/environments/test.rb
@@ -25,7 +25,7 @@ Rails.application.configure do
   config.action_dispatch.show_exceptions = false
 
   # Store uploaded files on the local file system in a temporary directory
-  config.active_storage.service = :test
+  # config.active_storage.service = :test
 
   # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr