+++ spec/spec_helper.rb
@@ -11,7 +11,6 @@ RSpec.configure do |config|
   config.filter_run_when_matching :focus
   config.disable_monkey_patching!
   config.warnings = false
-  config.default_formatter = 'doc' if config.files_to_run.one?
   config.profile_examples = 10
   config.order = :random
   Kernel.srand config.seed