+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## v1.0.0 - Unreleased
 
+### Features
+
+* Add Grape API Framework
+
 #### Technical Improvements
 
 * Add Rubocop linting